refactor: move PointMassForce class to new point_mass_force.py
